fix(core): add logger to static execute
In order to do that, needed to extract from core/logger/logger.go
the logic that depended on `core.` components. This was needed because
otherwise importing a logger would cause an import cycle.
