+++ mgc/core/config/config.go
@@ -3,12 +3,10 @@ package config
 import (
 	"context"
 	"fmt"
-	"log"
 	"os"
 	"path"
 
 	"magalu.cloud/core"
-	"magalu.cloud/core/logger"
 
 	"github.com/spf13/afero"
 	"github.com/spf13/viper"
@@ -47,8 +45,7 @@ func FromContext(ctx context.Context) *Config {
 func New() *Config {
 	dirname, err := core.BuildMGCPath()
 	if err != nil {
-		// TODO: when it's done, use logger instead
-		log.Println(err)
+		logger().Warnln(err)
 	}
 
 	c := &Config{}
@@ -73,7 +70,7 @@ func (c *Config) init(dirname string, fs afero.Fs) {
 }
 
 func (c *Config) BuiltInConfigs() (map[string]*core.Schema, error) {
-	loggerConfigSchema, err := logger.ConfigSchema()
+	loggerConfigSchema, err := loggerSchema()
 	if err != nil {
 		return nil, fmt.Errorf("unable to get logger config schema: %w", err)
 	}
